64bit upgrade

+++ kernel/idt.c
@@ -1,6 +1,6 @@
 #include "idt.h"
 
-void idt_set_entry(int n, unsigned int handler)
+void idt_set_entry(int n, unsigned long long handler)
 {
     idt[n].base_low  = handler & 0xFFFF;
     idt[n].base_high = (handler >> 16) & 0xFFFF;
@@ -12,7 +12,8 @@ void idt_set_entry(int n, unsigned int handler)
 void idt_init()
 {
     idt_ptr.limit = (sizeof(struct IDTEntry) * 256) - 1;
-    idt_ptr.base  = (unsigned int)&idt;
+   // TO:
+    idt_ptr.base = (unsigned long long)&idt;
 
     // Load IDT into CPU
     __asm__ volatile("lidt %0" : : "m"(idt_ptr));
